Add Project.reference_for_path
It mirrors the functionality of the method with
the same name in CocoaPod's Pod::Project class
(which after this patch will override it). The
differences are 1) CocoaPod's implementation is
constant time but this is linear wrt the number
of objects, 2) CocoaPod's implementation does not
work when you load a previously existing file.