[Scheme] Fix ReferencedContainer generation
Generate correct ReferencedContainer attribute
when the Xcode project has a non-empty
`projectDirPath`.

The ReferencedContainer attribute is actually
relative to the projectDirPath.

I ran into a problem where Xcode would generate
duplicate schemes of the hidden schemes that are
generated by CocoaPods because the
ReferencedContainer attribute that CocoaPods
generates was different from what Xcode generated
on an Xcode project that was generated by Cmake in
a certain way.