Remove unnecessary check for a Hash
Instances of `Hash` respond to `to_hash` (and just return themselves). We can
just check if an object responds to that method instead of checking for a `Hash`
and then falling back to the `to_hash` check.

The error message also now makes it a little bit clearer that anything that
responds to `to_hash` can be passed in.