Added support for saving schemes to workspaces.
When the scheme is saved in a workspace, and makes reference to targets
in projects that are not stored in the same directory as the
workspace, the container: fields must contain a path to the project
relative to the directory where the workspace is stored.  For this, we
need to set the bundlePath (and give the scheme name and shared flag
to build the scheme file name at the same time) before we add the
targets, so the XML is created correctly.  Unfortunately the XML code
is generated when we add objects, and the Xcodeproj objects are thrown
away immediately, keeping only the REXML::Element, so it's not
possible to do it when we save, and it is impossible to save different
workspace (in different directories) from the same scheme object!
This would have to be rewritten entirely.