Disable codesigning for iOS targets
Xcode 8 validates code-signing things as the first step when this
setting is configured at all. Since we want to not codesign for the Pods
project, we'll turn it off.

@segiddins we should probably do this inside CocoaPods instead and only
for Pod targets, but I couldn't figure out the syntax :grin:

```ruby
build_configuration.build_settings['CODE_SIGN_IDENTITY[sdk=iphoneos*]']
= ''
```

did not work for me.