Fix regression resulting in projects existing product refs being altered
Buddybuild uses xcodeproj against thousands of xcode projects, and there are a few we’ve noticed issues with as a result of this commit:

36ad554b Fixes issue CocoaPods/Xcodeproj#409

The problem is that xcodeproj was actually significantly altering the structure of a project if it already has a product_ref_group defined and it wasn’t part of the “Products” main group.

This would resultantly lead to fatal failures when we attempt to build the project.