PBXReferenceProxy: Add display_name attribute.
This commit fixes a bug that when a reference proxy doesn't have a
name, it's display name doesn't fall back to "ReferenceProxy" but
to the path.
This is manifested if you have a framework object in your project
and it's referenced (usually by a build file) in a build phase.