Fix crash when saving project.
Fixes a crash that happens when a project is saved, if that project had
an build phase exception set that's the child of a build phase that does
not have a name.
For example, the "Compile Sources" build phase does not have a name, and
could lead to this crash.